+++ packages/typespec-rust/src/codegen/models.ts
@@ -235,12 +235,17 @@ function emitModelDefinitions(module: rust.ModuleContainer, context: Context): h
 
       // TODO: omit skip_serializing_if if we need to send explicit JSON null
       // https://github.com/Azure/typespec-rust/issues/78
-      if (field.type.kind === 'option') {
+      if (field.flags & rust.ModelFieldFlags.ReadOnly) {
+        serdeParams.add('skip_serializing');
+      } else if (field.type.kind === 'option') {
         // optional literals need to skip serializing when it's None
         if ((field.type.type.kind !== 'enumValue' && field.type.type.kind !== 'literal') || field.optional) {
           serdeParams.add('skip_serializing_if = "Option::is_none"');
         }
-      } else if (model.visibility === 'pub') {
+      }
+
+      // TODO: this no longer seems to be required?
+      if (model.visibility === 'pub' && field.type.kind !== 'option') {
         // for public models, non-optional fields (e.g. Vec<T> in pageable responses) requires default.
         // crate models don't need this as those are used for spread params and the required params map
         // to the required fields in the struct.